# RentIt.Server/Models/Enums.cs
namespace RentIt.Server.Models;

public enum EquipmentCategory
{
    Electronics = 1,
    Tools = 2,
    Construction = 3,
    Garden = 4,
    SportsAndRecreation = 5,
}

public enum ItemStatus
{
    Available = 1,
    Rented = 2,
    Unavailable = 3,
}